cardioai/backend/requirements.txt
# ── Base Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# ── Database & ORM ──────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0  # Fallback for dev
alembic==1.13.1
pydantic[email]==2.7.3
pydantic-settings==2.2.1

# ── Security & Auth ─────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# ── ML Stack ────────────────────────────────────────────────
xgboost==2.0.3
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0

# ── Utilities ───────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
slowapi==0.1.9 # Rate limiting
